pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "mira-ai"
version = "2.2.0"
description = "Mira — AI 智能编程助手，支持 9 大 AI 提供商的 CLI + Web UI"
authors = [
    { name = "Your Name", email = "your.email@example.com" },
]
license = { file = "LICENSE" }
readme = "README.md"
requires-python = ">=3.8"
keywords = ["ai", "coding-assistant", "cli", "llm", "anthropic", "openai"]
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    # 核心 HTTP
    "requests>=2.31.0",
    "httpx>=0.25.0",
    # 搜索
    "ddgs>=9.0.0",
    # AI SDK
    "openai>=1.0.0",
    "anthropic>=0.18.0",
    "google-generativeai>=0.3.0",
    # Web UI（用于 --web 模式）
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.27.0",
    "websockets>=12.0",
    "aiofiles>=23.0.0",
    "pydantic>=2.0.0",
    # 可选功能
    "PyPDF2>=3.0.0",        # PDF 文件读取
]

[project.optional-dependencies]
web = [
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.27.0",
    "websockets>=12.0",
    "aiofiles>=23.0.0",
]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
]

[project.scripts]
mira  = "mira.main:main"
aicli = "mira.main:main"   # backward-compat alias

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
"mira.web" = ["static/*"]
